requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
fake-useragent==0.1.11
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
peewee==3.15.4
pygal==3.0.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
waitress==2.0.0
Werkzeug==2.0.3
zipp==3.7.0
